🐞 fix(VeterianDash, App, EmegencyContext): Bugs Se actualizo el canal para las notificaciones  Se modifico app y auth, para cargar el user antes y no tener mal intenciones
