+++ src/App.tsx
@@ -1,58 +1,57 @@
 import { useState } from "react";
-
 import { AuthProvider, useAuth } from "./contexts/AuthContext";
+import { EmergencyProvider } from "./contexts/EmergencyContext";
+
 import Login from "./components/auth/Login";
 import Register from "./components/auth/Register";
-import { EmergencyProvider } from "./contexts/EmergencyContext2";
-import ClientDashboard from "./components/dashboard/ClientDashboard";
 import LandingPage from "./components/LadingPage";
-
+import ClientDashboard from "./components/dashboard/ClientDashboard";
 import VeterinarianDashboard from "./components/dashboard/VeterinarianDashboard";
+
 function AppContent() {
   const [currentView, setCurrentView] = useState<
-    "landing" | "login" | "register" | "dashboard"
+    "landing" | "login" | "register"
   >("landing");
 
   const { user, loading } = useAuth();
+  console.log("Current user in App.tsx:", user);
 
-
-  console.log("token: ", localStorage.getItem("auth_token"));
-  console.log("user: ", user);
-
-
+  // ⏳ Mientras carga el usuario, mostramos un spinner
   if (loading) {
-    console.log("Loading user data...");
-
-    return <div className="flex items-center justify-center min-h-screen">Cargando...</div>;
+    return (
+      <div className="flex items-center justify-center min-h-screen">
+        Cargando...
+      </div>
+    );
+  } else if (user && localStorage.getItem("auth_token")) {
+    return user.role_id === "2" ? <ClientDashboard /> : <VeterinarianDashboard />;
   }
 
-  if (user && localStorage.getItem("auth_token")) {
-    console.log("User is logged in:", user.role_id);
-
-    return user.role_id === '2' ? <ClientDashboard />
-      : <VeterinarianDashboard /> // Aquí deberías tener un componente para el dashboard del veterinario>
+  else {
+    // Si no hay usuario, renderizamos las vistas de landing/login/register
+    switch (currentView) {
+      case "landing":
+        return <LandingPage onGetStarted={() => setCurrentView("login")} />;
+      case "login":
+        return (
+          <Login
+            onSwitchToRegister={() => setCurrentView("register")}
+            onBack={() => setCurrentView("landing")}
+          />
+        );
+      case "register":
+        return (
+          <Register
+            onSwitchToLogin={() => setCurrentView("login")}
+            onBack={() => setCurrentView("landing")}
+          />
+        );
+      default:
+        return <LandingPage onGetStarted={() => setCurrentView("login")} />;
+    }
   }
 
-  switch (currentView) {
-    case "landing":
-      return <LandingPage onGetStarted={() => setCurrentView("login")} />;
-    case "login":
-      return (
-        <Login
-          onSwitchToRegister={() => setCurrentView("register")}
-          onBack={() => setCurrentView("landing")}
-        />
-      );
-    case "register":
-      return (
-        <Register
-          onSwitchToLogin={() => setCurrentView("login")}
-          onBack={() => setCurrentView("landing")}
-        />
-      );
-    default:
-      return <LandingPage onGetStarted={() => setCurrentView("login")} />;
-  }
+
 }
 
 export default function App() {
