+++ src/main.tsx
@@ -21,6 +21,7 @@ createRoot(document.getElementById('root')!).render(
         },
       }}
     />
+
     <App />
   </StrictMode>,
 )
